# openapi.yaml
openapi: 3.0.0
info:
  title: Pictures API
  version: 1.0.0
paths:
  /pictures:
    get:
      summary: Получить список изображений
      responses:
        '200':
          description: Успешный запрос
          content:
            text/html:
              example: |
                <!DOCTYPE html>
                <html lang="ru">
                <head>
                    <meta charset="UTF-8">
                    <meta name="viewport" content="width=device-width, initial-scale=1.0">
                
                    <link rel="stylesheet" type="text/css" href="/style/Clear.css">
                    <link rel="stylesheet" type="text/css" href="/style/index.css">
                
                    <meta name="autor" content="Valkyrie (Егор Бордюков)">
                    <meta name="description" content="Свёрстанный макет на тему Онлайн галерея для тренировки">
                    <meta name="Keywords" content="онлайн,галерея,картины,верстка,макет сайта">
                
                    <title>Visual Realm</title>
                </head>
                <body id="body">
                    <div id="headder">
                        <div class="content-headder">
                            <img class="lodo-headder" src="./assets/image/logo.svg" alt="Visual Realm">
                            <div class="container-nav-button-headder">
                                <button class="nav-button-headder">Главная</button>
                                <button class="nav-button-headder">Выставка</button>
                                <button class="nav-button-headder">Информация</button>
                                <button class="nav-button-headder">Регистрация</button>
                            </div>
                        </div>
                    </div>
                    <div id="main">
                        <div class="banner">
                            <p id="text-main-banner">Онлайн галерея для публикации ваших работ</p>
                        </div>
                
                        <!-- Ширина изображения в блоке 400px
                        высота 264px -->
                
                        <div class="main-content"> 
                            <% data.forEach(item => { %>
                                <div class="block-content">
                                    <!-- вот изображение  -->
                                    <img class="block-img" src="<%= item.imageLink %>" alt="<%= item.name %>">
                                    <div class="meta-block">
                                        <div class="left-content">
                                            <!-- вот имя автора  -->
                                            <p class="nameAuthor"><%= item.author %></p>
                                            <div class="block-buttons">
                                                <button class="button-watch"><img src="./assets/image/button.svg" alt="Кнопка - посмотреть"></button>
                                                <button class="button-profile"><img src="./assets/image/button2.svg" alt="Кнопка - автор"></button>
                                            </div>
                                        </div>
                                        <!-- вот теги --> 
                                        <div class="tags">
                                            <p class="tag">#живопись</p>
                                            <p class="tag">#маслом</p>
                                            <p class="tag">#море</p>
                                            <p class="tag">#закат</p>
                                            <p class="tag">#пейзаж</p>
                                            <p class="tag">#иванов</p>
                                        </div>
                                    </div>   
                                </div>
                            <% }); %>
                            
                        </div>
                    </div>
                    <div class="footer">
                        <p class="nameAuthor" style="color: aliceblue; margin: 10px;">Valkyrie</p>
                        <a href="https://t.me/+FMChetgtqQBmMTFi">
                            <img src="./assets/image/telegram.svg" alt="telegram">
                        </a>
                    </div>
                </body>
                </html>
                
                
    post:
      summary: Добавить новое изображение
      requestBody:
        required: true
        content:
          application/json:
            example:
              name: SomeName
              author: SomeAuthor
              imageLink: someLink
             
      responses:
        '201':
          description: Изображение успешно добавлено
          content:
            application/json:
              example: 
                name: SomeName
                author: SomeAuthor
                imageLink: someLink
                
        '500':
          description: Ошибка добавления
                
  /pictures/{pictureId}:
    parameters:
      - name: pictureId
        in: path
        required: true
        description: ID изображения
        schema:
          type: integer
    put:
      summary: Обновить информацию об изображении по ID
      requestBody:
        required: true
        content:
          application/json:
            example: 
              name: SomeName
              author: SomeAuthor
              imageLink: someLink
      responses:
        '200':
          description: Изображение успешно обновлено
          content:
            application/json:
              example: 
                name: SomeName
                author: SomeAuthor
                imageLink: someLink
        '500':
          description: Ошибка обновления
    delete:
      summary: Удалить изображение по ID
      responses:
        '204':
          description: Изображение успешно удалено
        '500':
          description: Ошибка удаления